+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: mpaincha <mpaincha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/23 13:29:31 by mpaincha          #+#    #+#             */
-/*   Updated: 2015/12/02 16:34:34 by mpaincha         ###   ########.fr       */
+/*   Updated: 2015/12/03 10:47:03 by mpaincha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef	struct		s_list
 int					ft_atoi(const char *str);
 void				*ft_bzero(void *s, size_t n);
 int					ft_count_words(char const *s, char c);
-int					ft_intlen(int n);
+size_t				ft_intlen(int n);
 int					ft_isalnum(int c);
 int					ft_isalpha(int c);
 int					ft_isascii(int c);
